fix(security): move validatePublicServiceUrl out of frontend-shared module
validation.ts is imported by the React bundle for small helpers
(errMsg, isHttpUrl). Adding a top-level 'node:dns/promises' import
there broke the browser module graph in dev (vite) mode, which
rendered the SPA as a blank page and caused 8 e2e tests to fail
with element-not-found.

Move the DNS-dependent validatePublicServiceUrl to the new
src/server/url-safety.ts so only server code pulls in node:dns.
