+++ src/server/routes/settings.ts
@@ -3,14 +3,14 @@ import { createLastFmClient } from '@/core/clients/lastfm'
 import { createLidarrClient } from '@/core/clients/lidarr'
 import { createListenBrainzClient } from '@/core/clients/listenbrainz'
 import { sendWebhook } from '@/core/notifications'
-import { errMsg, validatePublicServiceUrl } from '@/core/validation'
+import { errMsg } from '@/core/validation'
 import { getUserConnections, updateUserConnections } from '@/db/queries/users'
 import type { Preferences } from '@/db/schema'
-
 import type { AppDependencies } from '@/server'
 import { resolveRequestMessages } from '@/server/locale'
 import { resolveAdmin } from '@/server/middleware/admin-guard'
 import type { HonoEnv } from '@/server/types'
+import { validatePublicServiceUrl } from '@/server/url-safety'
 
 const SECRET_FIELDS = [
   'lidarrApiKey',
